rename to fizzbuzzpy
